Making georouting federated
