+++ libs/blocks/global-navigation/utilities/utilities.js
@@ -1,5 +1,5 @@
 import {
-  getConfig, getMetadata, loadStyle, loadLana, decorateLinks, localizeLink,
+  getConfig, getMetadata, loadStyle, loadLana, decorateLinks, localizeLink, getFederatedContentRoot,
 } from '../../../utils/utils.js';
 import { processTrackingLabels } from '../../../martech/attributes.js';
 import { replaceText } from '../../../features/placeholders.js';
@@ -7,16 +7,6 @@ import { replaceText } from '../../../features/placeholders.js';
 loadLana();
 
 const FEDERAL_PATH_KEY = 'federal';
-
-// TODO when porting this to milo core, we should define this on config level
-// and allow consumers to add their own origins
-const allowedOrigins = [
-  'https://www.adobe.com',
-  'https://business.adobe.com',
-  'https://blog.adobe.com',
-  'https://milo.adobe.com',
-  'https://news.adobe.com',
-];
 
 export const selectors = {
   globalNav: '.global-navigation',
@@ -83,27 +73,6 @@ export function toFragment(htmlStrings, ...values) {
 
   return fragment;
 }
-
-// TODO we might eventually want to move this to the milo core utilities
-let federatedContentRoot;
-export const getFederatedContentRoot = () => {
-  if (federatedContentRoot) return federatedContentRoot;
-
-  const { origin } = window.location;
-
-  federatedContentRoot = allowedOrigins.some((o) => origin.replace('.stage', '') === o)
-    ? origin
-    : 'https://www.adobe.com';
-
-  if (origin.includes('localhost') || origin.includes('.hlx.')) {
-    // Akamai as proxy to avoid 401s, given AEM-EDS MS auth cross project limitations
-    federatedContentRoot = origin.includes('.hlx.live')
-      ? 'https://main--federal--adobecom.hlx.live'
-      : 'https://www.stage.adobe.com';
-  }
-
-  return federatedContentRoot;
-};
 
 // TODO we should match the akamai patterns /locale/federal/ at the start of the url
 // and make the check more strict.
